requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-phonenumber-field==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
Markdown==3.2.2
phonenumbers==8.12.8
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
